Amazing debugging.
This commit contains lots of debugging stuff, might be useful for later.

There were at least two completely distinct bugs I was hitting
simultaniously:

1. The final one I found was that assertNoEof should have been assert
   (not assertNot). This was AFTER finding the other bugs, dope!
2. The edge-case in c_charNext where there are no characters left was
   _totally_ broken. Finding that this was what was wrong took a _long_
   time, since I had never tested this.

I was convinced for the longest time that the problem was in the zstr
handling. It's not, but I'm going to be radically simplifying that
code in a second here.